set-up tournament add and model
