+++ py_backend/py_backend/settings.py
@@ -49,6 +49,7 @@ INSTALLED_APPS = [
 	'multiplayer',
 	'users',
 	'friends',
+	'tournaments',
 ]
 
 ASGI_APPLICATION = 'py_backend.asgi.application'
