+++ py_backend/requirements.txt
@@ -8,4 +8,6 @@ channels==4.0.0
 daphne>=4.0.0
 django-cors-headers==4.3.1
 pillow==10.2.0
+django-extensions==3.2.3
+django-bootstrap-v5
 # channels-redis==4.2.0
