+++ py_backend/tournaments/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class TournamentsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'tournaments'
